backend/requirements.txt
Django==2.2.14
djangorestframework==3.11.0
django-cors-headers==3.5.0
pytz==2020.1
sqlparse==0.3.1
